# openapi.yaml
openapi: 3.0.3
info:
  title: travel-api
  version: 1.0.0
  description: 旅行予約アプリ RESTful API

servers:
  - url: http://localhost:3000/v1
    description: ローカル開発

tags:
  - name: Health
  - name: Auth
  - name: Areas
  - name: Categories
  - name: Activities
  - name: Plans

paths:

  /health:
    get:
      tags: [Health]
      summary: ヘルスチェック
      operationId: getHealth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      tags: [Auth]
      summary: 新規登録
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: メールアドレス重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags: [Auth]
      summary: ログイン
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: アクセストークン更新
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [accessToken, refreshToken]
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          description: リフレッシュトークン無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      tags: [Auth]
      summary: ログアウト
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '204':
          description: ログアウト成功（No Content）

  /auth/me:
    get:
      tags: [Auth]
      summary: ログイン中ユーザー取得
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /areas:
    get:
      tags: [Areas]
      summary: エリア一覧（子エリア含む）
      operationId: getAreas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Area'

  /categories:
    get:
      tags: [Categories]
      summary: カテゴリ一覧（子カテゴリ含む）
      operationId: getCategories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

  /activities:
    get:
      tags: [Activities]
      summary: アクティビティ一覧・検索
      operationId: getActivities
      parameters:
        - name: keyword
          in: query
          description: タイトル・説明文をキーワード検索
          schema:
            type: string
        - name: area_id
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
        - name: available_date
          in: query
          description: 指定日に空きがあるアクティビティのみ返す（YYYY-MM-DD）
          schema:
            type: string
            format: date
            example: '2026-10-15'
        - name: min_price
          in: query
          schema:
            type: integer
            minimum: 0
        - name: max_price
          in: query
          schema:
            type: integer
            minimum: 0
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, price_asc, price_desc, rating, newest]
            default: popular
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: カーソルページング用（前レスポンスの pagination.nextCursor を渡す）
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivitySummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: クエリパラメータ不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /activities/{activityId}:
    get:
      tags: [Activities]
      summary: アクティビティ詳細
      description: 認証済みの場合 favorite.isFavorite を付与する。
      operationId: getActivityById
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ActivityDetail'
        '404':
          description: 見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /activities/{activityId}/plans:
    get:
      tags: [Activities]
      summary: アクティビティのプラン一覧
      operationId: getPlansByActivityId
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanWithMeetingPoints'
        '404':
          description: アクティビティが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /plans/{planId}:
    get:
      tags: [Plans]
      summary: プラン詳細
      operationId: getPlanById
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PlanDetail'
        '404':
          description: 見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    Error:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type:
          type: string
          example: https://api.example.com/errors/not-found
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        code:
          type: string
          example: NOT_FOUND
        detail:
          type: string
          example: リソースが見つかりません。

    Pagination:
      type: object
      required: [nextCursor, hasNext]
      properties:
        nextCursor:
          type: string
          nullable: true
        hasNext:
          type: boolean

    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true

    TokenResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'

    AreaChild:
      description: 子エリア（children を持たない）
      type: object
      required: [id, name, slug, sortOrder, isActive]
      properties:
        id:
          type: string
        parentId:
          type: string
          nullable: true
        name:
          type: string
        slug:
          type: string
        regionType:
          type: string
          nullable: true
        countryCode:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean

    Area:
      description: エリア（children 含む）
      allOf:
        - $ref: '#/components/schemas/AreaChild'
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: '#/components/schemas/AreaChild'

    CategoryChild:
      description: 子カテゴリ（children を持たない）
      type: object
      required: [id, name, slug, sortOrder, isActive]
      properties:
        id:
          type: string
        parentId:
          type: string
          nullable: true
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean

    Category:
      description: カテゴリ（children 含む）
      allOf:
        - $ref: '#/components/schemas/CategoryChild'
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: '#/components/schemas/CategoryChild'

    ActivityImage:
      type: object
      required: [id, url, imageType, sortOrder]
      properties:
        id:
          type: string
        url:
          type: string
        imageType:
          type: string
          enum: [main, gallery, map, plan, review]
        altText:
          type: string
          nullable: true
        sortOrder:
          type: integer

    PlanPrice:
      type: object
      required: [id, participantType, amount, currency]
      properties:
        id:
          type: string
        participantType:
          type: string
          enum: [adult, child, infant, senior]
        amount:
          type: integer
        currency:
          type: string
          example: JPY

    MeetingPoint:
      type: object
      required: [id, name, isDefault]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        accessInformation:
          type: string
          nullable: true
        meetingTime:
          type: string
          nullable: true
        isDefault:
          type: boolean

    PlanSchedule:
      type: object
      required: [id, sequence, title, isOptional]
      properties:
        id:
          type: string
        sequence:
          type: integer
        startTime:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        locationName:
          type: string
          nullable: true
        isOptional:
          type: boolean

    PlanBase:
      type: object
      required: [id, activityId, name, status, minParticipants]
      properties:
        id:
          type: string
        activityId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        minParticipants:
          type: integer
        maxParticipants:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, suspended]
        bookingDeadlineMinutes:
          type: integer
          nullable: true
        prices:
          type: array
          items:
            $ref: '#/components/schemas/PlanPrice'

    PlanWithMeetingPoints:
      description: プラン一覧で返すプラン（prices + meetingPoints 含む）
      allOf:
        - $ref: '#/components/schemas/PlanBase'
        - type: object
          properties:
            meetingPoints:
              type: array
              items:
                $ref: '#/components/schemas/MeetingPoint'

    PlanDetail:
      description: プラン詳細（schedules + activity 含む）
      allOf:
        - $ref: '#/components/schemas/PlanWithMeetingPoints'
        - type: object
          properties:
            schedules:
              type: array
              items:
                $ref: '#/components/schemas/PlanSchedule'
            activity:
              type: object
              properties:
                id:
                  type: string
                title:
                  type: string
                slug:
                  type: string
                images:
                  type: array
                  items:
                    $ref: '#/components/schemas/ActivityImage'
                  maxItems: 1
                area:
                  $ref: '#/components/schemas/AreaChild'
                category:
                  $ref: '#/components/schemas/CategoryChild'

    ActivityBase:
      type: object
      required: [id, title, slug, status, minPrice, maxPrice, currency, averageRating, reviewCount, bookingCount]
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, suspended, archived]
        minPrice:
          type: integer
        maxPrice:
          type: integer
        currency:
          type: string
          example: JPY
        averageRating:
          type: number
        reviewCount:
          type: integer
        bookingCount:
          type: integer
        area:
          $ref: '#/components/schemas/AreaChild'
        category:
          $ref: '#/components/schemas/CategoryChild'

    ActivitySummary:
      description: 一覧取得時のアクティビティ（メイン画像1枚）
      allOf:
        - $ref: '#/components/schemas/ActivityBase'
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: '#/components/schemas/ActivityImage'
              maxItems: 1

    ActivityDetail:
      description: 詳細取得時のアクティビティ（全画像・プラン・お気に入り状態含む）
      allOf:
        - $ref: '#/components/schemas/ActivityBase'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            highlights:
              type: array
              items:
                type: string
            images:
              type: array
              items:
                $ref: '#/components/schemas/ActivityImage'
            plans:
              type: array
              items:
                $ref: '#/components/schemas/PlanBase'
            favorite:
              type: object
              required: [isFavorite]
              properties:
                isFavorite:
                  type: boolean